✨ [Feat] (#7) QNA 양자택일 투표수 리스트 변환

+++ src/main/java/com/MtoM/MtoM/domain/qna/categroy/dao/QnaSelectResponse.java
@@ -10,7 +10,7 @@ public class QnaSelectResponse {
     private String select = "🧐양자택일"; // `🧐양자택일` 문자 고정
     private String title;
     private String createdAt;
-    private Long participants;
+    private int participants;
     private String userSelect; // 유저가 선택한 옵션 변환
     private List<VoteOptionResponse> options;
 }
