+++ src/main/java/com/MtoM/MtoM/domain/qna/categroy/dao/VoteOptionResponse.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class VoteOptionResponse {
     private String option1;
-    private int percentage1;
+    private Double percentage1;
     private String option2;
-    private int percentage2;
+    private Double percentage2;
 }
